voice-ai-agent-final/requirements.txt
fastapi
uvicorn
websockets
redis
openai
langdetect
pydantic
python-dotenv
